+++ src/lib/auth-components/guilds/join-guild/joined-contents/add-note.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import groupIcon from '$lib/assets/guild_group_icon_320.svg';
 	import { fade, scale } from 'svelte/transition';
 	import createIcon from '$lib/assets/create_guild_icon_320.svg';
