+++ src/lib/auth-components/learning-module/commenting/add-comment.svelte
@@ -1,0 +1,17 @@
+<script lang="ts">
+</script>
+
+<div class="">
+	<label>
+		<span class="text-main text-[14px] transition-all">Add Comment</span>
+		<textarea
+			class="outline-none w-full text-[14px] py-[11px] px-[20px] text-main bg-submain border-[1px] border-main rounded-[10px] transition-all"
+		/>
+	</label>
+</div>
+
+<button
+	class="transition-all active:bg-main/50 bg-main w-full rounded-[10px] text-[14px] font-semibold py-[10px] px-[2px] flex items-center justify-center text-submain"
+>
+	Add Comment
+</button>
